% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{all_numeric}
\alias{all_numeric}
\title{for a vector x,
check if all non-NA elements of x
can be converted to numeric}
\usage{
all_numeric(x)
}
\arguments{
\item{x}{A non-numeric vector}
}
\description{
for a vector x,
check if all non-NA elements of x
can be converted to numeric
}
