% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{get_gene_id}
\alias{get_gene_id}
\title{Get all gene IDs in a DESeqDataSet for a given gene name.}
\usage{
get_gene_id(gene_name, dds)
}
\arguments{
\item{gene_name}{A gene name}

\item{dds}{A DESeqDataSet}
}
\value{
A character vector
}
\description{
Get all gene IDs in a DESeqDataSet for a given gene name.
}
\examples{
get_gene_id("HBA1", T47D)
}
