% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{mean_sd_plot}
\alias{mean_sd_plot}
\title{Create a mean-sd plot
Make a scatterplot that shows for each gene its standard deviation versus mean.}
\usage{
mean_sd_plot(vsd)
}
\arguments{
\item{vsd}{A DESeqTransform object}
}
\value{
A ggplot object of the ggplot2 package that contains the mean-sd plot.
}
\description{
Create a mean-sd plot
Make a scatterplot that shows for each gene its standard deviation versus mean.
}
\examples{
\donttest{
library("DESeq2")
dds <- makeExampleDESeqDataSet(interceptMean=10, n=5000)
vsd <- vst(dds)
mean_sd_plot(vsd)
}
}
