% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other.R
\name{plot_biotypes}
\alias{plot_biotypes}
\title{Plot number of counts per sample and biotype}
\usage{
plot_biotypes(dds)
}
\arguments{
\item{dds}{A DESeqDataSet}
}
\value{
A ggplot object of the ggplot2 package.
}
\description{
Plot the total number of counts for each sample and the major classes of
ENSEMBL gene biotypes (protein coding, lncRNA, etc.)
}
\examples{
plot_biotypes(T47D)

}
