% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{plot_sample_boxplots}
\alias{plot_sample_boxplots}
\title{Plot one boxplot per sample}
\usage{
plot_sample_boxplots(se, se_assay = 1, group = NULL, ...)
}
\arguments{
\item{se}{A SummarizedExperiment}

\item{se_assay}{Name or the number of the assay to plot}

\item{group}{Either a variable (column name of \code{colData(se)}) by which to group samples or a
single number indicating how many samples to plot per group. Each group will be placed in a separate plot.
This is useful if the SummarizedExperiment contains more than 100 samples.}

\item{...}{Other arguments passed on to \code{aes()} inside \code{ggplot()}.}
}
\value{
A ggplot object of the ggplot2 package.

Either a list of ggplot objects of the ggplot2 package, or, if the list is length 1, its first element.
}
\description{
Plot one boxplot per sample
}
\details{
The function plots for each sample a bloxplot of its values. For more than 100 samples, individual boxplots
are hard to see and it's better to use the \code{group} parameter in that case, e.g. \code{group=100}.
}
\examples{
\donttest{
library("DESeq2")
vsd <- vst(T47D)
plot_sample_boxplots(vsd, group = "treatment")
plot_sample_boxplots(vsd, group = "mutation", color = treatment)
}
}
