% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{plot_sample_ecdfs}
\alias{plot_sample_ecdfs}
\title{Plot one ECDF per sample}
\usage{
plot_sample_ecdfs(
  se,
  se_assay = 1,
  n_steps = 100,
  line_alpha = 0.2,
  group = NULL,
  range = NULL
)
}
\arguments{
\item{se}{A SummarizedExperiment}

\item{se_assay}{Name or the number of the assay to plot}

\item{n_steps}{Number of points at which the ECDF is evaluated and plotted.}

\item{line_alpha}{Alpha value of the plotted lines.}

\item{group}{A variable (column name of \code{colData(se)}) by which to group samples.}

\item{range}{A numeric vector of length two specifying the interval of values to consider for the ECDFs.
Defaults to the maximal range of the values in the assay.}
}
\value{
A ggplot object of the ggplot2 package.
}
\description{
Plot one ECDF per sample
}
\details{
The function plots for each sample the ECDF of its values. This allows to identify outlier samples.
}
\examples{
\donttest{
library("DESeq2")
vsd <- vst(T47D)
plot_sample_ecdfs(vsd, group="treatment")
}
}
