% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAseqQC-package.R
\docType{package}
\name{RNAseqQC-package}
\alias{RNAseqQC}
\alias{RNAseqQC-package}
\title{RNAseqQC: Quality Control for RNA-Seq Data}
\description{
Functions for semi-automated quality control of bulk RNA-seq data.
}
\author{
\strong{Maintainer}: Frederik Ziebell \email{frederik.x.ziebell@gsk.com} (\href{https://orcid.org/0000-0003-3673-1721}{ORCID})

Other contributors:
\itemize{
  \item GlaxoSmithKline Research & Development Limited (GlaxoSmithKline Research & Development Limited; registered address: 980 Great West Road, Brentford, Middlesex TW8 9GS, United Kingdom) [copyright holder]
}

}
\keyword{internal}
