% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{T47D}
\alias{T47D}
\title{The T47D cell line data of RNA-seq experiment GSE89888}
\format{
A DESeqDataSet with 43576 rows (of genes) and 24 columns (of samples).
}
\source{
\doi{10.1101/2021.05.21.445138}
}
\usage{
T47D
}
\description{
The dataset contains the read counts of experiment GSE89888 in which T47D
cells with different mutation statuses were treated with E2 (estradiol) or
vehicle.
}
\keyword{datasets}
