% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{T47D_diff_testing}
\alias{T47D_diff_testing}
\title{Differential expression results corresponding to the T47D data set.}
\format{
A DESeqResults object with 36562 rows and 3 columns.
}
\source{
See the 'data' vignette on how to reproduce this object.
}
\usage{
T47D_diff_testing
}
\description{
Differential expression results corresponding to the T47D data set.
}
\keyword{datasets}
